+++ site/source/pages/Simulateurs/chômage-partiel/_config.ts
@@ -1,0 +1,47 @@
+import ChômagePartielComponent, { SeoExplanations } from '../ChômagePartiel'
+import { configChômagePartiel } from '../configs/chômagePartiel'
+import { config } from '../configs/config'
+import { SimulatorsDataParams } from '../configs/types'
+import ChômagePartielPreview from './images/ChômagePartielPreview.png'
+
+export function chômagePartielConfig({ t, sitePaths }: SimulatorsDataParams) {
+	return config({
+		id: 'chômage-partiel',
+		path: sitePaths.simulateurs['chômage-partiel'],
+		tracking: 'chomage_partiel',
+		pathId: 'simulateurs.chômage-partiel',
+		icône: '😷',
+		iframePath: 'simulateur-chomage-partiel',
+		meta: {
+			description: t(
+				'pages.simulateurs.chômage-partiel.meta.description',
+				"Calcul du revenu net pour l'employé et du reste à charge pour l'employeur après remboursement de l'Etat, en prenant en compte toutes les cotisations sociales."
+			),
+			ogDescription: t(
+				'pages.simulateurs.chômage-partiel.meta.ogDescription',
+				"Accédez à une première estimation en saisissant à partir d'un salaire brut. Vous pourrez ensuite personaliser votre situation (temps partiel, convention, etc). Prends en compte la totalité des cotisations, y compris celles spécifiques à l'indemnité (CSG-CRDS)."
+			),
+			ogTitle: t(
+				'pages.simulateurs.chômage-partiel.meta.ogTitle',
+				"Simulateur chômage partiel : découvrez l'impact sur le revenu net salarié et le coût total employeur."
+			),
+			title: t(
+				'pages.simulateurs.chômage-partiel.meta.titre',
+				"Calcul de l'indemnité chômage partiel : le simulateur Urssaf"
+			),
+			ogImage: ChômagePartielPreview,
+		},
+		shortName: t(
+			'pages.simulateurs.chômage-partiel.shortname',
+			'Chômage partiel'
+		),
+		title: t(
+			'pages.simulateurs.chômage-partiel.title',
+			"Simulateur du calcul de l'indemnité chômage partiel (Covid-19)"
+		),
+		nextSteps: ['salarié'],
+		simulation: configChômagePartiel,
+		component: ChômagePartielComponent,
+		seoExplanations: SeoExplanations,
+	} as const)
+}
