+++ site/source/pages/Simulateurs/impot-societe/_config.tsx
@@ -1,6 +1,6 @@
 import ISSimulation, { SeoExplanations } from '.'
 import { config } from '../configs/config'
-import { SimulatorsDataParams } from '../metadata'
+import { SimulatorsDataParams } from '../configs/types'
 import ISSimulationConfig from './_simulationConfig'
 
 export function impôtSociétéConfig({ t, sitePaths }: SimulatorsDataParams) {
